+++ src/data/Gallery.ts
@@ -17,4 +17,41 @@ export const workPictures: WorkImage[] = [
   { id: 7, title: 'Luxury Upgrade', image: '/Gallery/07.png', description: 'Premium material upgrade' },
   { id: 8, title: 'Custom Stitching', image: '/Gallery/08.png', description: 'Hand-stitched details' },
   { id: 9, title: 'Ergonomic Design', image: '/Gallery/09.png', description: 'Comfort-focused seating' },
+  { id: 10, title: 'Custom Car Interior', image: '/Gallery/10.png', description: 'Luxury car seat restoration' },
+  { id: 11, title: 'Truck Seat Upgrade', image: '/Gallery/11.png', description: 'Commercial truck seat modification' },
+  { id: 12, title: 'Boat Seat Project', image: '/Gallery/12.png', description: 'Marine seating solution' },
+  { id: 13, title: 'Motorcycle Custom', image: '/Gallery/13.png', description: 'Custom bike seat design' },
+  { id: 14, title: 'Classic Car Restoration', image: '/Gallery/14.png', description: 'Vintage car seat restoration' },
+  { id: 15, title: 'Modern Interior', image: '/Gallery/15.png', description: 'Contemporary seating design' },
+  { id: 16, title: 'Luxury Upgrade', image: '/Gallery/16.png', description: 'Premium material upgrade' },
+  { id: 17, title: 'Custom Stitching', image: '/Gallery/17.png', description: 'Hand-stitched details' },
+  { id: 18, title: 'Ergonomic Design', image: '/Gallery/18.png', description: 'Comfort-focused seating' },
+  { id: 19, title: 'Custom Car Interior', image: '/Gallery/19.png', description: 'Luxury car seat restoration' },
+  { id: 20, title: 'Truck Seat Upgrade', image: '/Gallery/20.png', description: 'Commercial truck seat modification' },
+  { id: 21, title: 'Boat Seat Project', image: '/Gallery/21.png', description: 'Marine seating solution' },
+  { id: 22, title: 'Motorcycle Custom', image: '/Gallery/22.png', description: 'Custom bike seat design' },
+  { id: 23, title: 'Classic Car Restoration', image: '/Gallery/23.png', description: 'Vintage car seat restoration' },
+  { id: 24, title: 'Modern Interior', image: '/Gallery/24.png', description: 'Contemporary seating design' },
+  { id: 25, title: 'Luxury Upgrade', image: '/Gallery/25.png', description: 'Premium material upgrade' },
+  { id: 26, title: 'Custom Stitching', image: '/Gallery/26.png', description: 'Hand-stitched details' },
+  { id: 27, title: 'Ergonomic Design', image: '/Gallery/27.png', description: 'Comfort-focused seating' },
+  { id: 28, title: 'Custom Car Interior', image: '/Gallery/28.png', description: 'Luxury car seat restoration' },
+  { id: 29, title: 'Truck Seat Upgrade', image: '/Gallery/29.png', description: 'Commercial truck seat modification' },
+  { id: 30, title: 'Boat Seat Project', image: '/Gallery/30.png', description: 'Marine seating solution' },
+  { id: 31, title: 'Motorcycle Custom', image: '/Gallery/31.png', description: 'Custom bike seat design' },
+  { id: 32, title: 'Classic Car Restoration', image: '/Gallery/32.png', description: 'Vintage car seat restoration' },
+  
+]; 
+export const workPicturesTruck: WorkImage[] = [
+  { id: 1, title: 'Truck Seat Customization', image: '/Gallery/Truckimages/truck01.jpg', description: 'Professional truck seat modification' },
+  { id: 2, title: 'Heavy Duty Truck Interior', image: '/Gallery/Truckimages/truck02.jpg', description: 'Commercial truck seat upgrade' },
+  { id: 3, title: 'Truck Seat Restoration', image: '/Gallery/Truckimages/truck03.jpg', description: 'Vintage truck seat restoration' },
+  { id: 4, title: 'Modern Truck Interior', image: '/Gallery/Truckimages/truck04.jpg', description: 'Contemporary truck seating design' },
+  { id: 5, title: 'Luxury Truck Upgrade', image: '/Gallery/Truckimages/truck05.jpg', description: 'Premium truck seat materials' },
+  { id: 6, title: 'Truck Seat Fabrication', image: '/Gallery/Truckimages/truck06.jpg', description: 'Custom truck seat fabrication' },
+  { id: 7, title: 'Truck Interior Design', image: '/Gallery/Truckimages/truck07.jpg', description: 'Professional truck interior design' },
+  { id: 8, title: 'Truck Seat Installation', image: '/Gallery/Truckimages/truck08.jpg', description: 'Expert truck seat installation' },
+  { id: 9, title: 'Truck Seat Maintenance', image: '/Gallery/Truckimages/truck09.jpg', description: 'Truck seat maintenance and care' },
+  { id: 10, title: 'Truck Seat Repair', image: '/Gallery/Truckimages/truck10.jpg', description: 'Truck seat repair services' },
+  { id: 11, title: 'Truck Seat Replacement', image: '/Gallery/Truckimages/truck11.jpg', description: 'Complete truck seat replacement' },
 ];
